--- server/index.js
@@ -1,6 +1,0 @@
-const app = require('./app');
-require('dotenv').config();
-
-
-//!Start Server
-app.start();
