--- server/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "express-mongo-sanitize": "^2.2.0",
-    "express-rate-limit": "^6.9.0",
-    "helmet": "^7.0.0",
-    "hpp": "^0.2.3",
-    "jsonwebtoken": "^9.0.1",
-    "mongoose": "^7.4.2",
-    "slugify": "^1.6.6",
-    "xss-clean": "^0.1.4"
-  },
-  "devDependencies": {
-    "dotenv": "^16.3.1",
-    "nodemon": "^3.0.1"
-  }
-}
