--- server/src/config/ConnectDB.js
@@ -1,13 +1,0 @@
-const mongoose = require('mongoose')
-
-const ConnectDB = async () => {
-    try {
-        mongoose.connect(process.env.DB_URI);
-        console.log('Database has been Connected!😊')
-    } catch (error) {
-        console.log('Database has been not Connected!😒')
-
-    }
-}
-
-module.exports = ConnectDB;
