+++ src/controllers/FoodController.js
@@ -1,5 +1,6 @@
 
-const { CreateFood, UpdateFood, DeleteFood, AllFoods, ReadFood, SearchByCategory, SearchByTitle } = require("../services/FoodService")
+const { CreateFood, UpdateFood, DeleteFood, AllFoods, ReadFood, SearchByCategory, SearchByTitle, FoodForPage } = require("../services/FoodService");
+const { VisitorCount } = require("../services/VisitorService");
 
 exports.CreateFood = async (req, res) => {
     const result = await CreateFood(req);
@@ -32,6 +33,12 @@ exports.AllFoods = async (req, res) => {
 
 };
 
+exports.FoodForPage = async (req, res) => {
+    const result = await FoodForPage(req);
+
+    res.status(200).json(result);
+
+};
 
 
 //!SEARCH
@@ -44,6 +51,15 @@ exports.SearchByCategory = async (req, res) => {
 
 exports.SearchByTitle = async (req, res) => {
     const result = await SearchByTitle(req);
+
+    res.status(200).json(result);
+};
+
+
+//!Visitor
+
+exports.VisitorCount = async (req, res) => {
+    const result = await VisitorCount(req);
 
     res.status(200).json(result);
 };
